requirements.txt
blinker==1.9.0
Bootstrap-Flask==2.5.0
click==8.2.1
dominate==2.9.1
Flask==3.1.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.1
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
